pyproject.toml
[project]
name = "pdf-splitter"
version = "0.2.0"
description = "PDF 分割アプリケーション (PySide6)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "PySide6>=6.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-qt>=4.5.0",
]

[tool.pytest.ini_options]
qt_api = "pyside6"
